refactor: [metrics] move chart rendering responsibility to graph-card-block
