+++ src/templates/graphs-card-block/components/chart/bar-chart/bar-chart.vue
@@ -33,3 +33,4 @@
     :id="`bar-chart-${props.chartData?.id}`"
   />
 </template>
+../../../helpers/c3-charts
