+++ src/templates/graphs-card-block/components/chart/line-chart/line-chart.vue
