+++ src/templates/graphs-card-block/components/chart/spline-chart/spline-chart.vue
@@ -33,3 +33,4 @@
     :id="`spline-chart-${props.chartData?.id}`"
   />
 </template>
+../../../helpers/c3-charts
