+++ src/templates/graphs-card-block/constants/chart-data.js
