+++ src/templates/graphs-card-block/constants/color-patterns.js
