+++ src/templates/graphs-card-block/helpers/c3-charts.js
@@ -1,17 +1,17 @@
 /* eslint-disable id-length */
 import {
-  RESET_COUNT,
-  MAX_COUNT,
-  MIN_COUNT,
+  BOTTOM_LEGEND_PADDING,
   C3_TYPES,
-  LABEL,
-  TO_FIXED_PERCENTAGE,
   DATA_VOLUME,
-  TO_FIXED_DATA_VOLUME,
+  LABEL,
+  MAX_COUNT,
+  MEAN_LINE_LABEL,
+  MIN_COUNT,
+  RESET_COUNT,
   SCREEN_SMALL_BREAKPOINT,
   SCREEN_XSMALL_BREAKPOINT,
-  BOTTOM_LEGEND_PADDING,
-  MEAN_LINE_LABEL
+  TO_FIXED_DATA_VOLUME,
+  TO_FIXED_PERCENTAGE
 } from '../constants/chart-data'
 import { LINE_PATTERNS, MEAN_LINE_PATTERN } from '../constants/color-patterns'
