+++ src/templates/graphs-card-block/index.vue
@@ -8,34 +8,48 @@
     <header class="flex w-full items-center justify-between gap-2">
       <span class="w-full gap-2 flex">
         <ChartOwner :chartOwner="chartOwner" />
-        <span class="font-medium overflow-ellipsis break-all line-clamp-1">{{ title }}</span>
+        <span class="font-medium overflow-ellipsis break-all line-clamp-1">{{ report.label }}</span>
       </span>
       <MoreOptionsMenu
-        :reportId="reportId"
+        :reportId="`${report.id}`"
         :clipboardWrite="clipboardWrite"
       />
     </header>
     <div class="flex h-full flex-col gap-6 flex-auto">
       <div class="flex flex-col">
         <span class="break-words text-sm text-color-secondary font-normal line-height-1 py-3.5">
-          {{ description }}
+          {{ report.description }}
         </span>
         <AggregationInfo
-          :aggregationType="aggregationType"
-          :variationType="variationType"
-          :variationValue="variationValue"
-          :displayTag="displayTag"
+          :aggregationType="report.aggregationType"
+          variationType="positive"
+          variationValue="10.2%"
+          :displayTag="true"
         />
       </div>
       <section class="flex-auto">
-        <slot name="chart" />
+        <component
+          v-if="report.resultQuery?.length && showChart"
+          :is="propToComponent[`${report.type}-chart`]"
+          :chartData="report"
+          :resultChart="report.resultQuery"
+          :hasMeanLineTotal="report.showMeanLine"
+          :hasMeanLineSeries="report.showMeanLinePerSeries"
+        />
+        <Skeleton
+          v-else
+          class="w-full h-full"
+        />
       </section>
     </div>
   </div>
 </template>
 
 <script setup>
-  import { computed } from 'vue'
+  import { useHelpCenterStore } from '@/stores/help-center'
+  import { storeToRefs } from 'pinia'
+  import Skeleton from 'primevue/skeleton'
+  import { computed, defineAsyncComponent, onMounted, onUnmounted, ref, watch } from 'vue'
   import AggregationInfo from './components/aggregation-info.vue'
   import ChartOwner from './components/chart-owner.vue'
   import MoreOptionsMenu from './components/more-options-menu.vue'
@@ -50,54 +64,16 @@
         return ['azion', 'account', 'user'].includes(value)
       }
     },
-    title: {
-      type: String,
+    report: {
+      type: Object,
       required: true
     },
-    description: {
-      type: String,
-      required: true
-    },
-    cols: {
-      type: Number,
-      default: 6,
-      validator(value) {
-        return [4, 6, 8, 12].includes(value)
-      }
-    },
-    aggregationType: {
-      type: String,
-      default: 'sum',
-      validator(value) {
-        return ['sum', 'avg'].includes(value)
-      }
-    },
-    displayTag: {
-      type: Boolean,
-      default: true
-    },
-    variationType: {
-      type: String,
-      default: 'none',
-      validator(value) {
-        return [
-          'none',
-          'positive',
-          'negative',
-          'positive-inverse',
-          'negative-inverse',
-          'not-compare'
-        ].includes(value)
-      }
-    },
-    variationValue: { type: String, default: '' },
-    reportId: String,
     clipboardWrite: Function
   })
 
   const cardColumns = computed(() => {
     // For some reason, template strings do not work here
-    switch (props.cols) {
+    switch (props.report.columns) {
       case 4:
         return 'lg:col-span-4'
       case 8:
@@ -107,5 +83,43 @@
       default:
         return 'lg:col-span-6'
     }
+  })
+
+  const propToComponent = {
+    'bar-chart': defineAsyncComponent(() => import('./components/chart/bar-chart/bar-chart')),
+    'line-chart': defineAsyncComponent(() => import('./components/chart/line-chart/line-chart')),
+    'spline-chart': defineAsyncComponent(() =>
+      import('./components/chart/spline-chart/spline-chart')
+    )
+  }
+
+  const { getStatus } = storeToRefs(useHelpCenterStore())
+
+  const showChart = ref(true)
+
+  const reRenderChart = () => {
+    /*
+     * This approach was used to re-render the chart after screen resize and help center toggling.
+     * The nextTick was not sufficient to resolve the latter.
+     * HelpCenter takes 300ms to animate. To garantee the proper re-render, 350ms was used.
+     */
+    const timeout = 350
+
+    showChart.value = false
+    setTimeout(() => {
+      showChart.value = true
+    }, timeout)
+  }
+
+  onMounted(() => {
+    window.addEventListener('resize', reRenderChart)
+  })
+
+  onUnmounted(() => {
+    window.removeEventListener('resize', reRenderChart)
+  })
+
+  watch(getStatus, () => {
+    reRenderChart()
   })
 </script>
